+++ Opc.Ua.CloudLib.Client/Models/QueryModel.cs
