+++ UA-CloudLibrary.sln
@@ -17,7 +17,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleConsoleClient", "SampleConsoleClient\SampleConsoleClient.csproj", "{E38A75A7-8CAC-4233-9D4D-0C855E815ABF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Opc.Ua.CloudLib.Client", "UACloudLibraryGraphQLClient\UACloudLibClientLibrary\Opc.Ua.CloudLib.Client.csproj", "{8353C72D-E29A-40BE-A3EF-C59330FC270C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Opc.Ua.CloudLib.Client", "Opc.Ua.CloudLib.Client\Opc.Ua.CloudLib.Client.csproj", "{8353C72D-E29A-40BE-A3EF-C59330FC270C}"
 EndProject
 Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{C7282E18-BCDA-48F6-99FD-EC7142A0397A}"
 EndProject
