+++ version.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://raw.githubusercontent.com/dotnet/Nerdbank.GitVersioning/master/src/NerdBank.GitVersioning/version.schema.json",
-  "version": "1.0-beta",
+  "version": "0.1.0-beta",
   "nuGetPackageVersion": {
     "semVer": 2.0
   },
